Renamed NewNav into NavbarItems for a better description
